requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.55.3
httpx>=0.23.0, <1.0.0
python-dotenv==1.0.1
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.24.1
fpdf2==2.7.8
sqlalchemy==2.0.29
psycopg2-binary==2.9.9